+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Adding `xan input --vcf, --gtf`.
 * Adding `xan search -R/--replace` & `xan search --replacement-column`.
 * Adding `xan rename -S/--slugify`.
+* Adding moonblade function `sum`.
 
 *Fixes*